Update README.md
This portfolio website is created with HTML, CSS, and JavaScript to showcase Sourav Kundu's skills and projects. It features a sleek, responsive design with sections like Home, Skills, Projects, and Contact. HTML structures the content, CSS styles it for a modern look, and JavaScript adds smooth interactions. The site highlights Sourav's expertise in areas such as programming, machine learning, and web development. It offers a user-friendly experience, making it easy for potential employers or collaborators to explore Sourav's work and get in touch directly